add config and restructure
